services/booking-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
prometheus-client==0.19.0
motor==3.3.2
pymongo==4.6.1
redis==5.0.1
python-dotenv==1.0.0
stripe==11.5.0
aiohttp==3.10.11
asyncpg==0.29.0
# Structured logging
structlog==24.1.0
python-json-logger==2.0.7
# Distributed tracing
opentelemetry-sdk==1.24.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-instrumentation-httpx==0.45b0
